+++ packages/reactive_ble_platform_interface/lib/src/model/qualified_characteristic.dart
@@ -26,7 +26,7 @@ class QualifiedCharacteristic {
       "$runtimeType(characteristicId: $characteristicId, serviceId: $serviceId, deviceId: $deviceId)";
 
   @override
-  int get hashCode =>
+  BigInt get hashCode =>
       (((17 * 37) + characteristicId.hashCode) * 37 + serviceId.hashCode) * 37 +
       deviceId.hashCode;
